Add files via upload
For this machine learning problem, I am trying to find how accurate the model detects the nature of the comments found in the imdb dataset. How accurate can it be able to detect whether the comments are positive or negative. To tackle this problem, I am going to be using the IMDB dataset which consists movie reviews and sentiment annotations.

This machine learning probblem can be classified under Binary classification using the IMDB movie reviews data to see if the reviews on the website are positive or negative. Binary classification is used as this dataset invloves 2 classes in which is positive or negative. The data will be obtained from keras dataset.